further cleanup #113
