+++ app/components/EntitiesOverTime/ChartTimelineCategories.js
