+++ app/components/EntitiesOverTime/PlotHintContent.js
