+++ app/containers/EntityList/messages.js
@@ -50,10 +50,6 @@ export default defineMessages({
     id: 'app.containers.EntityList.filterFormError',
     defaultMessage: 'Errors',
   },
-  highlightCategory: {
-    id: 'app.containers.EntitiesCategories.highlightCategory',
-    defaultMessage: 'Highlight by {categoryName}',
-  },
   viewOptionList: {
     id: 'app.containers.EntityList.viewOptionList',
     defaultMessage: 'List',
